requirements.txt
# Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis
redis[hiredis]==5.2.0

# Auth
python-jose[cryptography]==3.3.0
httpx==0.28.0
google-auth==2.37.0

# Validation
pydantic==2.10.0
pydantic-settings==2.7.0
email-validator==2.2.0

# Rate limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
factory-boy==3.3.1
aiosqlite==0.20.0
